Removed the wait time as a parameter for the station rating and increased the importance of the units of waiting cargo instead.
I did this to make large capacity vehicles roughly as efficient as multiple small capacity ones. In reality cargo trains are few and long. Nobody sends out 100 short trains. We send out few long ones. The game is currently punishing that. A station at a coal mine can have a miserable rating despite not even having enough cargo waiting to fill the one train servicing it when it returns. That is stupid.