Get rid of some problematic asset and hopefully fix the very very rare problem.
Even if it still occurs, it won't affect anything. The game tries several times to connect cities.
Worse case a city remains unconnected.
Also introduce a sensible max for the search nodes. About a million is enough and will almost
never be exceeded, and even if, the game will find another way to hook up the city later.